#20 #21 move delete_all_players and delete_all_rooms into corresponding small modules
